Have removed shtml usage as github doesnt allow server side includes (ssi).
This may be a problem in the future, so we need to see how we continue
with these pages without this ability.

In files with extension .shtml, on many systems, a comment tag followed
by "#include" would include html code, like in c++, which is an easy way
to reuse html code, like for menus for instance, in all the webpages.
Github pages cannot work with this, so we will need to see how we
progress, if this is the route we take.